Bug 430390 - Missing call to resolve path parameters in method-path annotation for CreationFactory annotation
resolving path parameters in @path annotations of both the class and the
methods in a more consistent way. This is done by moving the logic to
resolve paramters into the private method resolvePathParameters (which
originally only handled class paramters)

The pattern now is applied for CreationFactory, queryCapability, and
Dialogs.

Also added a basic unit test for the ServiceProviderFactory class.

Note: Unlike CreationFactory & QueryCapability, The URI Dialogs seem to
be independent of the class path-annotation. Only the method
@path-annotation is relevant. This code difference is maintained in this
patch.

Change-Id: Iaadb3c4a63b2dd344ca410a8b49f6ef49cfa7251
Signed-off-by: Jad <jad@kth.se>
Signed-off-by: Andrew Berezovskyi <andriib@kth.se>