Bug 517298 - Generated createResourceShape method on a Resource class produces wrong URIs
Most methods in the application (including createResourceShape) that
need to define the resource's URI are dependent on the servletURI
(localhost:8080/adaptor-webapp/services), which is not easily available.
Only the publicURI (localhost:8080/adaptor-webapp) is available.


Introduced a getter/setter method in OSLC4JUtils to set the servletPath
value. This value is then use when requesting the ServletURI, which is a
concatenation of the PublicURI and the servletPath.

Also, modified ServletListener so it register the baseURI and servletURI
in the OSLC4JUtils. Application no longer need to ask ServletListener
for URIs. OSLC4JUtils is the reference.

Plus, no longer assume "resources" for the servletPath. read instead the
url-pattern from web.xml file.

Finally, Base/Servlet pathes are now consistently obtained from the
OSLC4JUtils methods, instead of asking ServletListener of the
application. This would allow classes to work with client applications,
where no listener is available. (yet the application still need to
register the pathes to OSLC4JUTils)



Change-Id: I3c4cc6c10abba1a75fd27e68bbd5f7b5bc2b9e66
Signed-off-by: Jad <jad@kth.se>