Bug 517339 - The URI of a resouceShape from a GET request to OslcResourceShapeResource does not match the requested URI
The GET method in OslcResourceShapeResource, calls resolveURI with the
paramter "includePath" set to false. This excludes the servletPath from
the resolvedURI, leading to the wrong shape URI.

Adding the servletPath to the baseURI.

Note: this is the only call to resolveURI with "includePath" set to
false. I wonder if it makes more sense for this method to return the
path including the servletPath. Is there any use of the contextPath
only?

Change-Id: I035a7f923334e33457fed0e8f2839a4080c6c239