Bug 517339 - URI of a resouceShape does not match the requested URI
The GET method in OslcResourceShapeResource, calls resolveURI with the
parameter "includePath" set to false. This excludes the servletPath from
the resolvedURI, leading to the wrong shape URI.

Instead, resolveUri is deprecated in favour of two new methods:

* resolveFullUri
* resolveServletUri

Both of these methods try to use paths set on OSLC4JUtils, otherwise
they obtain them from the HttpServletRequest object.

New tests are merged with the methods from the ResolveURITest.

Change-Id: I59865f2df98bb66eb679e4b882041cdaa4f36637
Signed-off-by: Andrew Berezovskyi <andriib@kth.se>